Mejora en la ruta /delete de carrito

+++ src/controllers/carritoController.js
@@ -4,6 +4,7 @@ import {productsAPI} from '../apis/productos';
 import { Gmail } from '../services/gmail';
 import {SmsService} from '../services/twilio'
 import { logger } from '../utils/logs';
+import { UserAPI } from '../apis/user';
 let carrito =[
     {
         id:1, 
@@ -111,33 +112,46 @@ class Carrito{
           productId,
           {stock:totalstock} 
         );
-        console.log(updatedProduct)
         res.json({ msg: 'Producto agregado con exito', cart: updatedCart });
       }
     
       async deleteProduct(req, res) {
         const user = req.user;
-        const cart = await carritoAPI.getCarrito(user._id);
+        const {id} =req.params
+        const cart = await carritoAPI.getCarrito(user[0]._id);
     
         const { productId, productAmount } = req.body;
-    
+        const userId = user[0]._id;
+        if (id != userId)
+          return res.status(400).json({ msg: 'User id not found' });
+        
         if (!productId || !productAmount)
           return res.status(400).json({ msg: 'Invalid body parameters' });
     
         const product = await productsAPI.getProducts(productId);
-    
+       
         if (!product.length)
-          return res.status(400).json({ msg: 'product not found' });
+          return res.status(400).json({ msg: 'Product not found' });
     
         if (parseInt(productAmount) < 0)
           return res.status(400).json({ msg: 'Invalid amount' });
     
+        const proAmount = cart.productos[0].amount;
+        if(parseInt(productAmount) > proAmount)
+          return res.status(400).json({ msg: `La cantidad que ingresa supera la cantidad actual(${proAmount}) del producto en el carrito` });
+
         const updatedCart = await carritoAPI.deleteProudct(
           cart._id,
           productId,
           parseInt(productAmount)
         );
-        res.json({ msg: 'Product deleted', cart: updatedCart });
+        let totalstock = product[0].stock + productAmount
+        let stock = product[0].stock
+        let updatedProduct = await productsAPI.updateProduct(
+          productId,
+          {stock:totalstock} 
+        );
+        res.json({ msg: 'Product deleted' ,cart:updatedCart});
       }
 
       async comprarProduct(req, res) {
